added icon theme and asking the user before logging out

+++ install.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/bash
 sudo -v
 sudo pacman -Syu
-sudo pacman -Sy hyprland nvim waybar fish awww kitty tmux nemo rofi network-manager-applet hyprpolkitagent base-devel hyprlock gtk3 gtk4 qt6ct superfile zoxide 
+sudo pacman -Sy hyprland nvim waybar fish awww kitty tmux nemo rofi network-manager-applet hyprpolkitagent base-devel hyprlock gtk3 gtk4 qt6ct superfile zoxide pavucontrol papirus-icon-theme
 echo ""
 read -p "Install yay? <y/N> " prompt
 if [[ $prompt == "y" || $prompt == "Y" || $prompt == "yes" || $prompt == "Yes" ]]
@@ -20,6 +20,10 @@ cp -r ./.config/kitty ~/.config/
 cp -r ./.config/hypr ~/.config/
 cp -r ./.config/waybar ~/.config/
 cp -r ./.config/nvim ~/.config/
+cp -r ./Pictures ~/Pictures
 
-pkill -u $USER
+read -p "Logout? <y/N> " prompt
+if [[ $prompt == "y" || $prompt == "Y" || $prompt == "yes" || $prompt == "Yes" ]]
+	pkill -u $USER
+fi
